refactor(tui): split repl_app.rs into repl_app + repl_render modules
Extract all pure rendering/drawing functions and text utilities into
repl_render.rs (1344 lines). repl_app.rs retains state management,
event handling, slash commands, OAuth threads, entry points, and tests
(2487 lines).

repl_app.rs was 3816 lines — a single file handling REPL UI, event
loop, modal management, rendering, and state. The split separates
the what-to-draw (repl_render) from the when-and-why (repl_app).
