+++ crates/acrawl-cli/src/tui/mod.rs
@@ -10,6 +10,7 @@ pub mod grouped_model_list;
 pub mod modal;
 pub mod model_list;
 pub mod model_modal;
+pub(super) mod repl_render;
 pub mod repl_app;
 pub mod tool_panel;
